+++ app/src/main/java/com/mobiquity/androidunittests/calculator/input/operator/SubtractionOperator.java
@@ -1,6 +1,5 @@
 package com.mobiquity.androidunittests.calculator.input.operator;
 
-import com.mobiquity.androidunittests.calculator.input.Input;
 import com.mobiquity.androidunittests.calculator.input.InputType;
 
 public class SubtractionOperator extends Operator {
@@ -10,7 +9,7 @@ public class SubtractionOperator extends Operator {
     }
 
     @Override
-    public int execute(int param1, int param2) {
+    public double execute(double param1, double param2) {
         return subtract(param1, param2);
     }
 
@@ -24,7 +23,7 @@ public class SubtractionOperator extends Operator {
         return true;
     }
 
-    private int subtract(int minuend, int subtrahend) {
+    private double subtract(double minuend, double subtrahend) {
         return minuend - subtrahend;
     }
